requirements.txt
absl-py==0.2.0
admesh==0.98.8
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astor==0.6.2
audioread==2.1.5
backcall==0.1.0
beautifulsoup4==4.6.1
bleach==1.5.0
bokeh==0.12.16
Brlapi==0.6.6
catkin-pkg==0.4.9
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
command-not-found==0.3
cryptography==2.1.4
cssselect==1.0.3
cupshelpers==1.0
cycler==0.10.0
Cython==0.28.4
decorator==4.3.0
defer==1.0.6
distro-info==0.18
dlib==19.10.0
dnspython==1.15.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
eventlet==0.24.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-Cors==3.0.4
Flask-SocketIO==3.0.2
Flask-WTF==0.14.2
gast==0.2.0
google-speech==1.0.16
greenlet==0.4.15
grpcio==1.11.0
gunicorn==19.9.0
h5py==2.7.1
html5lib==0.9999999
httplib2==0.9.2
idna==2.6
imageai==2.0.2
imageio==2.3.0
imutils==0.4.6
inflect==1.0.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
itsdangerous==0.24
jedi==0.12.0
jieba3k==0.35.1
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.6
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
leveldb==0.1
librosa==0.6.0
llvmlite==0.23.0
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
metrics==0.3.3
mistune==0.8.3
monotonic==1.5
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
newspaper3k==0.2.6
nltk==3.3
nose==1.3.7
notebook==5.4.1
numba==0.38.0
numexpr==2.6.4
numpy==1.14.3
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
opencv-contrib-python==3.4.0.12
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
pathlib2==2.3.2
pathspec==0.5.5
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
pipenv==2018.7.1
pipreqs==0.4.9
plotly==2.6.0
progressbar==2.5
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycairo==1.16.2
pycocotools==2.0.0
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pygal==2.4.0
Pygments==2.2.0
pygobject==3.26.1
pyimagesearch==0.4
pymacaroons==0.13.0
PyNaCl==1.1.2
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
pyparsing==2.3.0
PyQt5==5.10.1
pyrealsense2==2.16.2.306
pyRFC3339==1.0
PySocks==1.6.8
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
python-engineio==2.2.0
python-gflags==1.5.1
python-socketio==2.0.0
python-speech-features==0.6
pytz==2018.3
PyWavelets==0.5.2
pyxdg==0.25
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
reportlab==3.4.0
requests==2.18.4
requests-file==1.4.3
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
resampy==0.2.0
resources==0.0.1
rosdep==0.13.0
rosdistro==0.7.0
rosinstall==0.7.8
rosinstall-generator==0.1.14
rospkg==1.1.7
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
scour==0.36
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.13.2
SimpleWebSocketServer==0.1.0
sip==4.19.8
six==1.11.0
sklearn==0.0
SoundFile==0.10.2
staty==0.5.0
sympy==1.2
system-service==0.3
systemd-python==234
tables==3.4.2
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-gpu==1.8.0
tensorlayer==1.8.4
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tflearn==0.3.2
tldextract==2.2.0
torch==0.4.0
torchvision==0.2.1
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
trimesh==2.31.26
tweepy==3.6.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
vaderSentiment==3.2.1
vcstools==0.1.40
virtualenv==16.0.0
virtualenv-clone==0.3.0
wadllib==1.3.2
wcwidth==0.1.7
web-cache==1.0.2
webencodings==0.5
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wstool==0.1.17
WTForm==1.0
WTForms==2.2.1
xkit==0.0.0
xlib==0.21
yarg==0.1.9
zope.interface==4.3.2
